+++ source/MUWorldRegistry.h
@@ -18,14 +18,14 @@
 
 + (MUWorldRegistry *) defaultRegistry;
 
-- (void) insertObject: (MUWorld *) world inWorldsAtIndex: (unsigned) worldIndex;
-- (void) removeObjectFromWorldsAtIndex: (unsigned) worldIndex;
+- (void) insertObject: (MUWorld *) world inWorldsAtIndex: (NSUInteger) worldIndex;
+- (void) removeObjectFromWorldsAtIndex: (NSUInteger) worldIndex;
 
 - (NSUInteger) count;
 - (NSUInteger) indexOfWorld: (MUWorld *) world;
 - (void) removeWorld: (MUWorld *) world;
 - (void) replaceWorld: (MUWorld *) oldWorld withWorld: (MUWorld *) newWorld;
-- (MUWorld *) worldAtIndex: (unsigned) worldIndex;
+- (MUWorld *) worldAtIndex: (NSUInteger) worldIndex;
 - (MUWorld *) worldForUniqueIdentifier: (NSString *) identifier;
 
 @end
