+++ source/infrastructure/MUReadBuffer.m
@@ -95,7 +95,7 @@ NSString *MUReadBufferDidProvideStringNotification = @"MUReadBufferDidProvideStr
   return [self length] == 0;
 }
 
-- (unsigned) length
+- (NSUInteger) length
 {
   return [dataBuffer length];
 }
